Setup test environement, and requests tests

+++ services/requests/taskRequests.ts
@@ -1,9 +1,9 @@
 "use server";
 
-import { DOE } from "@/types/common";
+import { DOE, JSONType } from "@/types/common";
 import { AuthUser, ProjectModel } from "@/types/models";
 import { projectSchema, partialProjectSchema } from "@/helpers/validators";
-import { supabase } from "@/helpers/supabase";
+import { supabaseClient } from "@/helpers/supabase";
 import { convertToProjectModel } from "@/helpers/converters";
 import {
   canCreateProject,
@@ -16,26 +16,27 @@ import { filterQuery, PaginatedData } from "@/helpers/query";
 /* --------------------------------------------------------
    CREATE PROJECT
 ---------------------------------------------------------*/
-export async function requestCreateProject(formData: FormData, authUser: AuthUser): Promise<DOE<ProjectModel>> {
+export async function requestCreateProject(jsonData: JSONType, authUser: AuthUser): Promise<DOE<ProjectModel>> {
   try {
-    if (!authUser) return { data: null, error: { message: "Unauthorized" } };
+    if (!authUser) return { data: null, error: { message: "Unauthorized", code: "403" } };
 
     const projectsCount = await requestGetUserProjectsCount(authUser.id, authUser);
     if (!canCreateProject(authUser, projectsCount)) {
-      return { data: null, error: { message: "Unauthorized" } };
+      return { data: null, error: { message: "Unauthorized, maximum allowed project reached", code: "403" } };
     }
 
-    const newProjectData = Object.fromEntries(formData.entries());
-    const parsed = projectSchema.safeParse(newProjectData);
+    // const newProjectData = Object.fromEntries(formData.entries());
+    const newProjectData = jsonData;
+    const parsed = projectSchema.omit({ user_id: true, id: true, created_at: true, updated_at: true }).safeParse(newProjectData);
 
     if (!parsed.success) {
       const errors = parsed.error.flatten().fieldErrors;
       return { data: null, error: { message: "Validation failed", errors } };
     }
 
-    const { data, error } = await supabase
+    const { data, error } = await supabaseClient
       .from("projects")
-      .insert({ ...parsed.data, user_id: authUser.id })
+      .insert({ ...parsed.data, user_id: authUser.id, updated_at: new Date().toISOString(), created_at: new Date().toISOString() })
       .select()
       .single();
 
@@ -56,19 +57,18 @@ export async function requestCreateProject(formData: FormData, authUser: AuthUse
 ---------------------------------------------------------*/
 export async function requestGetProject(projectId: string, authUser: AuthUser): Promise<DOE<ProjectModel>> {
   try {
-    const { data, error } = await supabase
+    const { data, error } = await supabaseClient
       .from("projects")
       .select("*")
       .eq("id", projectId)
       .single();
 
     if (error) return { data: null, error: { message: error.message } };
-    if (!data) return { data: null, error: { message: "Project not found" } };
 
     const project = convertToProjectModel(data);
 
     if (project && !canGetProject(authUser, project)) {
-      return { data: null, error: { message: "You don't have access to this project" } };
+      return { data: null, error: { message: "You don't have access to this project", code: "403" } };
     }
 
     return { data: project, error: null };
@@ -84,13 +84,10 @@ export async function requestGetProject(projectId: string, authUser: AuthUser): 
 export async function requestGetUserProjects(userId: string, authUser: AuthUser): Promise<DOE<ProjectModel[]>> {
   try {
     if (userId !== authUser?.id) {
-      return { data: null, error: { message: "Unauthorized operation" } };
+      return { data: null, error: { message: "Unauthorized operation", code: "403" } };
     }
 
-    const { data, error } = await supabase
-      .from("projects")
-      .select("*")
-      .eq("user_id", userId);
+    const { data, error } = await supabaseClient.from("projects").select("*").eq("user_id", userId);
 
     if (error) return { data: null, error: { message: error.message } };
 
@@ -110,29 +107,29 @@ export async function requestGetUserProjects(userId: string, authUser: AuthUser)
 ---------------------------------------------------------*/
 export async function requestGetUserProjectsCount(userId: string, authUser: AuthUser): Promise<number> {
   try {
-    if (userId !== authUser?.id) return 0;
+    if (userId !== authUser?.id) return -1;
 
-    const { count, error } = await supabase
+    const { count, error } = await supabaseClient
       .from("projects")
       .select("*", { count: "exact", head: true })
       .eq("user_id", userId);
 
     if (error) {
       console.error("[getProjectsCount] Error fetching count:", error);
-      return 0;
+      return -1;
     }
 
-    return count ?? 0;
+    return count ?? -1;
   } catch (error: any) {
     console.error("[getProjectsCount] Unexpected error:", error);
-    return 0;
+    return -1;
   }
 }
 
 /* --------------------------------------------------------
    UPDATE PROJECT
 ---------------------------------------------------------*/
-export async function requestUpdateProject(projectId: string, formData: FormData, authUser: AuthUser): Promise<DOE<ProjectModel>> {
+export async function requestUpdateProject(projectId: string, jsonData: JSONType, authUser: AuthUser): Promise<DOE<ProjectModel>> {
   try {
     const projectResp = await requestGetProject(projectId, authUser);
     if (!projectResp.data) return { data: null, error: { message: "Project not found" } };
@@ -140,21 +137,21 @@ export async function requestUpdateProject(projectId: string, formData: FormData
     const project = projectResp.data;
 
     if (!canUpdateProject(authUser, project)) {
-      return { data: null, error: { message: "You don't have access to this project" } };
+      return { data: null, error: { message: "You don't have access to this project", code: "403" } };
     }
 
-    const updates = Object.fromEntries(formData.entries());
-    const parsed = partialProjectSchema.safeParse(updates);
+    const updates = jsonData;
+    const parsed = partialProjectSchema.omit({updated_at: true, created_at: true, id: true}).safeParse(updates);
 
     if (!parsed.success) {
       const errors = parsed.error.flatten().fieldErrors;
       return { data: null, error: { message: "Validation failed", errors } };
     }
 
-    const { data, error } = await supabase
+    const { data, error } = await supabaseClient
       .from("projects")
       .update({ ...parsed.data, updated_at: new Date().toISOString() })
-      .eq("id", projectId)
+      .eq("id", projectId).eq("user_id", authUser?.id)
       .select()
       .single();
 
@@ -173,7 +170,7 @@ export async function requestUpdateProject(projectId: string, formData: FormData
 export async function requestDeleteProject(projectId: string, authUser: AuthUser): Promise<DOE<boolean>> {
   try {
     const projectResp = await requestGetProject(projectId, authUser);
-    if (!projectResp.data) return { data: false, error: { message: "Project not found" } };
+    if (!projectResp.data) return { data: false, error: { message: "Project not found", code: "403" } };
 
     const project = projectResp.data;
 
@@ -181,10 +178,10 @@ export async function requestDeleteProject(projectId: string, authUser: AuthUser
       return { data: false, error: { message: "You don't have access to this project" } };
     }
 
-    const { error } = await supabase
+    const { error } = await supabaseClient
       .from("projects")
       .delete()
-      .eq("id", projectId);
+      .eq("id", projectId).eq("user_id", authUser?.id);
 
     if (error) return { data: false, error: { message: error.message } };
 
@@ -199,7 +196,7 @@ export async function requestDeleteProject(projectId: string, authUser: AuthUser
 export async function requestFilteredProfiles(queryParams: URLSearchParams, authUser: AuthUser, defaultSelect: string = "*"):
   Promise<DOE<PaginatedData<ProjectModel>>> {
   try {
-    const filteredData = await filterQuery(supabase.from("projects"), queryParams, defaultSelect);
+    const filteredData = await filterQuery(supabaseClient.from("projects"), queryParams, defaultSelect);
     const profiles: ProjectModel[] = filteredData.data.map(entry => convertToProjectModel(entry)).
       filter(entry => !!entry).filter(entry => canGetProject(authUser, entry));
     return {
